+++ packages/fsproductdetail/src/index.ts
@@ -1,5 +1,5 @@
 export { default as ProductDetail } from './components/ProductDetail';
-
+export { default as ProductDetailGhost } from './components/ProductDetailGhost';
 export {
   default as withProductDetailData,
   WithProductDetailProps,